refactor: embeddings
