+++ cortex-js/src/infrastructure/commanders/embeddings.command.ts
@@ -7,7 +7,6 @@ import {
 import ora from 'ora';
 import { ModelsUsecases } from '@/usecases/models/models.usecases';
 import { PSCliUsecases } from './usecases/ps.cli.usecases';
-import { ChatCliUsecases } from './usecases/chat.cli.usecases';
 import { inspect } from 'util';
 import { ModelStat } from './types/model-stat.interface';
 import { CortexUsecases } from '@/usecases/cortex/cortex.usecases';
@@ -30,8 +29,6 @@ interface EmbeddingCommandOptions {
 })
 export class EmbeddingCommand extends BaseCommand {
   constructor(
-    private readonly chatCliUsecases: ChatCliUsecases,
-    private readonly modelsUsecases: ModelsUsecases,
     private readonly psCliUsecases: PSCliUsecases,
     private readonly inquirerService: InquirerService,
     readonly cortexUsecases: CortexUsecases,
@@ -42,30 +39,33 @@ export class EmbeddingCommand extends BaseCommand {
     passedParams: string[],
     options: EmbeddingCommandOptions,
   ): Promise<void> {
-    let modelId = passedParams[0];
+    let model = passedParams[0];
     const checkingSpinner = ora('Checking model...').start();
     // First attempt to get message from input or options
     let input: string | string[] = options.input ?? passedParams.splice(1);
 
     // Check for model existing
-    if (!modelId || !(await this.modelsUsecases.findOne(modelId))) {
+    if (!model || !(await this.cortex.models.retrieve(model))) {
       // Model ID is not provided
       // first input might be message input
       input = passedParams ?? options.input;
       // If model ID is not provided, prompt user to select from running models
       const models = await this.psCliUsecases.getModels();
       if (models.length === 1) {
-        modelId = models[0].modelId;
+        model = models[0].modelId;
       } else if (models.length > 0) {
-        modelId = await this.modelInquiry(models);
+        model = await this.modelInquiry(models);
       } else {
         checkingSpinner.fail('Model ID is required');
         process.exit(1);
       }
     }
     checkingSpinner.succeed(`Model found`);
-    return this.chatCliUsecases
-      .embeddings(modelId, input)
+    return this.cortex.embeddings
+      .create({
+        input,
+        model,
+      })
       .then((res) =>
         inspect(res, { showHidden: false, depth: null, colors: true }),
       )
